Improve deployment variables (#94)
* Missing attributes

Since DEPLOYEMENT_TIME was decided to be a double, it should have the
required attributes to describe its meaning.

* The coordinates should be double as well

It doesn't make sense to use String for deployment lat/lon. If I'm not
missing anything here, it should also have some attributes such as used
for the other lat/lon.

* Updating example to reflect changes

* fix: Left some attributes without updating

Thanks to @callumrollo who noticed that.