Conventions as mandatory (#91)
* Conventions should be mandatory

Conventions is a fundamental metadata used by applications defining what
to expect on this file. There is no reason to don't be mandatory.

Thanks to @vturpin who noticed this.

* Update CF version example

This can't be a fixed value since we shall allow the user to incorporate
other conventions. Therefore, no consequences on which version is the
example, so let's update this.

* Clean to avoid confusion

On "OG1.0 General conventions" it is defined that OG follows CF and
ACDD-1.3, so the extra text here doesn't make sense and just create
confusion.