Adding UUID in the sensors' table
This table is outdated with this current proposal, but I'm writing here
anyways so we don't forget the UUID field. It was decided in this PR to
add this field as suggested so we can guide the standard to be used but
without imposing more requirements for the data providers.