Updating the document to reflect sensors as groups
This needs some improvements such as some guidance on what are groups.
Not everyone will be familiar with that. If we use groups in other parts
of this dataset we might define groups in one place only. Anyways, the
sensors session deserves a better text but I'll postpone that for now.