Sensors as groups (#103)
Major contributions from @emmerbodc and @vturpin.

List of commits:

* Sensors as groups

* Fixing vocabularies

Thanks to @emmerbodc who noticed those errors!

* No need for dimension SENSORS

By using groups, each sensor is moved to a group, thus no need for
dimension or variable sensor.

* Adding link between measurement and sensor

* Missing identifiers' level in the vocabulary attributes

Thanks to @emmerbodc who noticed that!

* Improving some vocabularies and long names

Thanks to @emmerbodc!

* Adding UUID in the sensors' table

This table is outdated with this current proposal, but I'm writing here
anyways so we don't forget the UUID field. It was decided in this PR to
add this field as suggested so we can guide the standard to be used but
without imposing more requirements for the data providers.

* Fixing platform vocabularies

Thanks to @ emmerbodc!

* A history document

I'm creating this file as a place holder to record the contributions and
history of the standard. This will be an important appendix to the
official document to support the list of contributors.

We shall be transparent and fair giving proper credits to all contributors.

* With sensors as groups we don't need PARAMETER_SENSOR anymore

Thanks to @vturpin that raised that on #105.

* We don't need N_SENSOR anymore

Sensors will be stored as groups.

* Updating the document to reflect sensors as groups

This needs some improvements such as some guidance on what are groups.
Not everyone will be familiar with that. If we use groups in other parts
of this dataset we might define groups in one place only. Anyways, the
sensors session deserves a better text but I'll postpone that for now.