Adding CF-1.8 Single Trajectory Template
I've attempted to generate a CF-1.8 compliant nc/cdl following the guidelines in the OG-1.0 docs. This cdl is meant to show how the dimensions would be setup given the OG-1.0 spec. Note that the metadata here are not to OG-1.0 spec and are not nearly as complete the existing example cdl (https://github.com/OceanGlidersCommunity/OG-format-user-manual/blob/main/src/sp041_20191205T1757.cdl).  I've added PARAM_STRLEN, which is just a requirement in CF-1.8 to have the character length. I selected 16 chars, based on an Argo sample data file and the parameter name vocabulary doc looks like all names are within this limit. If the file doesn't contain more than one trajectory it does not need the Trajectory dimension. I've added a TRAJ_STRLEN dimension which is similarly just the character length of the trajectory name. I'm not sure how leaving the trajectory dimension out will impact aggregation, I expect that the dimension would be added upon aggregation. If you plan to aggregate you may want to consider testing. Standardizing the TRAJ_STRLEN number or following a different discrete geometry example may be beneficial for aggregation.