Change agency into institution.adoc (#154)
* Change agency into institution.adoc

"Institution" is used by the three orginial format. While agency isn't. To keep it as simple as possible and facilitate use of the new format, we suggest here to move from agency to institution in the OG1.0.

Agency comes from OceanOPS vocabularies only. There is no strong reason not to use institution as it is already a concept used by the key DACs. 
We did not pay attention to that details until now.

* institution's into institutions'.adoc

Correct typo

* Institutions instead of institution.adoc

correcting typo

* typo.adoc

* typo.adoc