Adding GPS variables
It is interesting that it is not clear here it all these attributes are
mandatory. And the most important, it is not clear if the values are
examples or fixed values, i.e. everyone should repeat exactly those
values? For instance, if there is no QC, should the
'ancillary_variables' be ignored?