Create folder with exemplar - create minimal.cdl
Here I propose the minimal .cdl

I've removed everything that is not mandatory.
This is align with the status of the format the 14th of February 2024 with downgrading of PARAMETER and SENSOR informations.

Impact should be low as the agreement on what is mandatory in OG.1.0 is pretty weak. But a final check is necessary before release.