Create metadataFormat.adoc
Following issue #75 and anticipating on OG1.0 format delivery. I am proposing here a metadata file that will agregate all metadata into a json file, easy to read and very usefull for monitoring (amongst other thing).

This is a practice adopted by Argo with very good results in terms of monitoring.

This is to be discussed amongst each other and not to be tackle before the iugc2024.